Don't delete non-existing directories. Check for null before closing database connection.

+++ Machine.Migrations/Machine.Migrations/Services/Impl/WorkingDirectoryManager.cs
@@ -46,7 +46,10 @@ namespace Machine.Migrations.Services.Impl
 
     public void Destroy()
     {
-      _fileSystem.RemoveDirectory(this.WorkingDirectory);
+      if (_fileSystem.IsDirectory(this.WorkingDirectory))
+      {
+        _fileSystem.RemoveDirectory(this.WorkingDirectory);
+      } 
     }
     #endregion
   }
